requirements.txt
# Core dependencies
alpaca-py>=0.21.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0

# Technical analysis
pandas-ta>=0.4.67b0

# Sentiment analysis
transformers>=4.36.0
torch>=2.1.0

# Web dashboard
flask>=3.0.0

# Utilities
schedule>=1.2.0
pydantic>=2.5.0
pytz>=2024.1

# Logging and monitoring
structlog>=24.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
isort>=5.13.0
